+++ NEWS.md
@@ -1,3 +1,15 @@
+# ordbetareg v0.7.0
+
+- Allow intercepts to receive separate priors from main effects (permitting zero-ing out intercepts).
+- Changed log-likelihood calculation to match `brms` documentation and permit point-wise 
+`loo` calculation.
+- Added animated density plots for continuous responses.
+- Improved plot formatting and added theme and label options.
+- Updated vignette to include information about `glmmTMB` as an alternative for estimation
+and new plot functions.
+- Added distribution `rordbeta` and `dordbeta` functions.
+- Removed legacy `0 + Intercept` function code. Now accepts any kind of formula.
+
 # ordbetareg v0.5.0
 
 - Fixed bug disabling extra priors for phi regression/modeling.